Parse Fhir types with hapi jsonParser
